+++ test/Benchmarks/Instrumentation/AspNetCoreInstrumentationBenchmarks.cs
@@ -17,30 +17,70 @@
 #if !NETFRAMEWORK
 using BenchmarkDotNet.Attributes;
 using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Logging;
 using OpenTelemetry;
 using OpenTelemetry.Metrics;
 using OpenTelemetry.Trace;
 
 /*
-// * Summary *
+BenchmarkDotNet=v0.13.5, OS=Windows 11 (10.0.23424.1000)
+Intel Core i7-9700 CPU 3.00GHz, 1 CPU, 8 logical and 8 physical cores
+.NET SDK=7.0.203
+  [Host]     : .NET 7.0.5 (7.0.523.17405), X64 RyuJIT AVX2
+  DefaultJob : .NET 7.0.5 (7.0.523.17405), X64 RyuJIT AVX2
 
-BenchmarkDotNet=v0.13.3, OS=Windows 10 (10.0.19045.2604)
-Intel Core i7-4790 CPU 3.60GHz (Haswell), 1 CPU, 8 logical and 4 physical cores
-.NET SDK=7.0.103
-  [Host] : .NET 7.0.3 (7.0.323.6910), X64 RyuJIT AVX2
 
-Job=InProcess  Toolchain=InProcessEmitToolchain
+|                     Method | EnableInstrumentation |     Mean |   Error |  StdDev |   Gen0 | Allocated |
+|--------------------------- |---------------------- |---------:|--------:|--------:|-------:|----------:|
+| GetRequestForAspNetCoreApp |                  None | 226.8 us | 4.00 us | 3.74 us |      - |   2.45 KB |
+| GetRequestForAspNetCoreApp |                Traces | 235.2 us | 4.44 us | 4.15 us | 0.4883 |   3.59 KB |
+| GetRequestForAspNetCoreApp |               Metrics | 229.1 us | 4.44 us | 4.36 us |      - |   2.92 KB |
+| GetRequestForAspNetCoreApp |       Traces, Metrics | 230.6 us | 4.54 us | 5.23 us | 0.4883 |   3.66 KB |
 
-|                                      Method |     Mean |   Error |  StdDev |   Gen0 | Allocated |
-|-------------------------------------------- |---------:|--------:|--------:|-------:|----------:|
-|                 UninstrumentedAspNetCoreApp | 149.4 us | 2.94 us | 2.75 us | 0.4883 |   2.54 KB |
-| InstrumentedAspNetCoreAppWithDefaultOptions | 171.9 us | 2.65 us | 2.48 us | 0.7324 |   3.79 KB |
+Allocation details for .NET 7:
+
+// Traces
+* Activity creation + `Activity.Start()` = 416 B
+* Casting of the struct `Microsoft.Extensions.Primitives.StringValues` to `IEnumerable<string>` by `HttpRequestHeaderValuesGetter`
+  - `TraceContextPropagator.Extract` = 24 B
+  - `BaggageContextPropagator.Extract` = 24 B
+  - `request.Headers["User-Agent"].FirstOrDefault()` = 24 B
+* String creation for `HttpRequest.HostString.Host` = 40 B
+* `Activity.TagsLinkedList` (this is allocated on the first Activity.SetTag call) = 40 B
+* Boxing of `Port` number when adding it as a tag = 24 B
+* String creation in `GetUri` method for adding the http url tag = 66 B
+* Setting `Baggage` (Setting AsyncLocal values causes allocation)
+  - `BaggageHolder` creation = 24 B
+  - `System.Threading.AsyncLocalValueMap.TwoElementAsyncLocalValueMap` = 48 B
+  - `System.Threading.ExecutionContext` = 40 B
+* `DiagNode<KeyValuePair<System.String, System.Object>>`
+  - This is allocated eight times for the eight tags that are added = 8 * 40 = 320 B
+* `Activity.Stop()` trying to set `Activity.Current` (This happens because of setting another AsyncLocal variable which is `Baggage`
+  - System.Threading.AsyncLocalValueMap.OneElementAsyncLocalValueMap = 32 B
+  - System.Threading.ExecutionContext = 40 B
+
+Baseline = 2.45 KB
+With Traces = 2.45 + (1162 / 1024) = 2.45 + 1.14 = 3.59 KB
+
+
+// Metrics
+* Activity creation + `Activity.Start()` = 416 B
+* Boxing of `Port` number when adding it as a tag = 24 B
+* String creation for `HttpRequest.HostString.Host` = 40 B
+
+Baseline = 2.45 KB
+With Metrics = 2.45 + (416 + 40 + 24) / 1024 = 2.45 + 0.47 = 2.92 KB
+
+// With Traces and Metrics
+
+Baseline = 2.45 KB
+With Traces and Metrics = Baseline + With Traces + (With Metrics - (Activity creation + `Acitivity.Stop()`)) (they use the same activity)
+                        = 2.45 + (1162 + 64) / 1024 = 2.45 + 1.2 = 3.55 KB (~3.56 KB)
 */
 
 namespace Benchmarks.Instrumentation
 {
-    [InProcess]
     public class AspNetCoreInstrumentationBenchmarks
     {
         private HttpClient httpClient;
@@ -48,55 +88,102 @@ namespace Benchmarks.Instrumentation
         private TracerProvider tracerProvider;
         private MeterProvider meterProvider;
 
-        [GlobalSetup(Target = nameof(UninstrumentedAspNetCoreApp))]
-        public void UninstrumentedAspNetCoreAppGlobalSetup()
+        [Flags]
+        public enum EnableInstrumentationOption
         {
-            this.StartWebApplication();
-            this.httpClient = new HttpClient();
+            /// <summary>
+            /// Instrumentation is not enabled for any signal.
+            /// </summary>
+            None = 0,
+
+            /// <summary>
+            /// Instrumentation is enbled only for Traces.
+            /// </summary>
+            Traces = 1,
+
+            /// <summary>
+            /// Instrumentation is enbled only for Metrics.
+            /// </summary>
+            Metrics = 2,
         }
 
-        [GlobalSetup(Target = nameof(InstrumentedAspNetCoreAppWithDefaultOptions))]
-        public void InstrumentedAspNetCoreAppWithDefaultOptionsGlobalSetup()
+        [Params(0, 1, 2, 3)]
+        public EnableInstrumentationOption EnableInstrumentation { get; set; }
+
+        [GlobalSetup(Target = nameof(GetRequestForAspNetCoreApp))]
+        public void GetRequestForAspNetCoreAppGlobalSetup()
         {
-            this.StartWebApplication();
-            this.httpClient = new HttpClient();
+            if (this.EnableInstrumentation == EnableInstrumentationOption.None)
+            {
+                this.StartWebApplication();
+                this.httpClient = new HttpClient();
+            }
+            else if (this.EnableInstrumentation == EnableInstrumentationOption.Traces)
+            {
+                this.StartWebApplication();
+                this.httpClient = new HttpClient();
 
-            this.tracerProvider = Sdk.CreateTracerProviderBuilder()
-                .AddAspNetCoreInstrumentation()
-                .Build();
+                this.tracerProvider = Sdk.CreateTracerProviderBuilder()
+                    .AddAspNetCoreInstrumentation()
+                    .Build();
+            }
+            else if (this.EnableInstrumentation == EnableInstrumentationOption.Metrics)
+            {
+                this.StartWebApplication();
+                this.httpClient = new HttpClient();
 
-            var exportedItems = new List<Metric>();
-            this.meterProvider = Sdk.CreateMeterProviderBuilder()
-                .AddAspNetCoreInstrumentation()
-                .AddInMemoryExporter(exportedItems)
-                .Build();
+                this.meterProvider = Sdk.CreateMeterProviderBuilder()
+                    .AddAspNetCoreInstrumentation()
+                    .Build();
+            }
+            else if (this.EnableInstrumentation.HasFlag(EnableInstrumentationOption.Traces) &&
+                this.EnableInstrumentation.HasFlag(EnableInstrumentationOption.Metrics))
+            {
+                this.StartWebApplication();
+                this.httpClient = new HttpClient();
+
+                this.tracerProvider = Sdk.CreateTracerProviderBuilder()
+                    .AddAspNetCoreInstrumentation()
+                    .Build();
+
+                this.meterProvider = Sdk.CreateMeterProviderBuilder()
+                    .AddAspNetCoreInstrumentation()
+                    .Build();
+            }
         }
 
-        [GlobalCleanup(Target = nameof(UninstrumentedAspNetCoreApp))]
-        public async Task GlobalCleanupUninstrumentedAspNetCoreAppAsync()
+        [GlobalCleanup(Target = nameof(GetRequestForAspNetCoreApp))]
+        public void GetRequestForAspNetCoreAppGlobalCleanup()
         {
-            this.httpClient.Dispose();
-            await this.app.DisposeAsync().ConfigureAwait(false);
-        }
-
-        [GlobalCleanup(Target = nameof(InstrumentedAspNetCoreAppWithDefaultOptions))]
-        public async Task GlobalCleanupInstrumentedAspNetCoreAppWithDefaultOptionsAsync()
-        {
-            this.httpClient.Dispose();
-            await this.app.DisposeAsync().ConfigureAwait(false);
-            this.tracerProvider.Dispose();
-            this.meterProvider.Dispose();
+            if (this.EnableInstrumentation == EnableInstrumentationOption.None)
+            {
+                this.httpClient.Dispose();
+                this.app.DisposeAsync().GetAwaiter().GetResult();
+            }
+            else if (this.EnableInstrumentation == EnableInstrumentationOption.Traces)
+            {
+                this.httpClient.Dispose();
+                this.app.DisposeAsync().GetAwaiter().GetResult();
+                this.tracerProvider.Dispose();
+            }
+            else if (this.EnableInstrumentation == EnableInstrumentationOption.Metrics)
+            {
+                this.httpClient.Dispose();
+                this.app.DisposeAsync().GetAwaiter().GetResult();
+                this.meterProvider.Dispose();
+            }
+            else if (this.EnableInstrumentation.HasFlag(EnableInstrumentationOption.Traces) &&
+                this.EnableInstrumentation.HasFlag(EnableInstrumentationOption.Metrics))
+            {
+                this.httpClient.Dispose();
+                this.app.DisposeAsync().GetAwaiter().GetResult();
+                this.tracerProvider.Dispose();
+                this.meterProvider.Dispose();
+            }
         }
 
         [Benchmark]
-        public async Task UninstrumentedAspNetCoreApp()
-        {
-            var httpResponse = await this.httpClient.GetAsync("http://localhost:5000").ConfigureAwait(false);
-            httpResponse.EnsureSuccessStatusCode();
-        }
-
-        [Benchmark]
-        public async Task InstrumentedAspNetCoreAppWithDefaultOptions()
+        public async Task GetRequestForAspNetCoreApp()
         {
             var httpResponse = await this.httpClient.GetAsync("http://localhost:5000").ConfigureAwait(false);
             httpResponse.EnsureSuccessStatusCode();
@@ -107,7 +194,7 @@ namespace Benchmarks.Instrumentation
             var builder = WebApplication.CreateBuilder();
             builder.Logging.ClearProviders();
             var app = builder.Build();
-            app.MapGet("/", () => $"Hello World!");
+            app.MapGet("/", async context => await context.Response.WriteAsync($"Hello World!"));
             app.RunAsync();
 
             this.app = app;