+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+* Revert version of `Google.Protobuf` to `3.19.4` (see
+  [#4201](https://github.com/open-telemetry/opentelemetry-dotnet/pull/4201)).
+  This also reintroduces the `System.Reflection.Emit.Lightweight` dependency.
+  ([#4407](https://github.com/open-telemetry/opentelemetry-dotnet/pull/4407))
+
 ## 1.5.0-alpha.2
 
 Released 2023-Mar-31